Fix flickering in TempoInstance dropdown
Previously, the dropdown was unselected until the list of TempoResources
was loaded. With this change, the current instance based on the query
params is pre-selected with a slightly inaccurate TempoResourceSelectOption
(e.g. list of tenants is inaccurate), which will be replaced with
the correct TempoResourceSelectOption once the API call finished.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>