Use TimeRangeProvider to manipulate time range
Previously, every change to the query resulted in two Tempo API calls,
because the <TimeRangeProvider> component of the <PersesDashboardWrapper>
received a new object for the `timeRange` prop on every render,
resulting in two re-renders:

(1) due to a change in props and
(2) due to a change in the first hook (localTimeRange):
```
useEffect(() => {
  setLocalTimeRange(timeRange);
}, [timeRange, refreshCounter]);
```

This change moves the <TimeRangeProvider> up and uses its exposed
methods to update and refresh the current time range. This results in a
single Tempo API call per search.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>